+++ packages/eslint-config-airbnb-base/rules/es6.js
@@ -124,7 +124,6 @@ module.exports = {
 
     // suggest using Reflect methods where applicable
     // http://eslint.org/docs/rules/prefer-reflect
-    // TODO: enable?
     'prefer-reflect': 'off',
 
     // use rest parameters instead of arguments
@@ -165,6 +164,8 @@ module.exports = {
 
     // enforce spacing around the * in yield* expressions
     // http://eslint.org/docs/rules/yield-star-spacing
-    'yield-star-spacing': ['error', 'after']
+    'yield-star-spacing': ['error', 'after'],
+
+    'disallowParenthesesAroundArrowParam': true
   }
 };